requirements.txt
flask
mysql-connector-python
python-dotenv
pandas
gunicorn
flask-login
werkzeug
cryptography
flask-talisman
flask-wtf
glob2
